Use fewer bits to store the lookup table, promote as needed instead, which is easy

+++ board.h
@@ -33,7 +33,7 @@ class Board{
 	static const uint64_t xybits[36];       // xybits[i] = (1ull << xytobit[i])
 	static const uint64_t winmaps[32];      // the bit patterns for the 32 win conditions
 	static const uint64_t flipquad[512];    // lookup table to mirror a quadrant
-	static const uint64_t * lookup3to2;     // convert base 3 for one line of 6 to base 2, used in hashing
+	static const uint16_t * lookup3to2;     // convert base 3 for one line of 6 to base 2, used in hashing
 	static const int      scoremap[6];      // how many points a given line with how many pieces is worth
 
 	uint64_t sides[3]; // sides[0] = sides[1] | sides[2];
@@ -248,10 +248,10 @@ private:
 		uint64_t h = 0;
 		uint64_t w = sides[1] << 9;
 		uint64_t b = sides[2];
-		h |= lookup3to2[((w & (0x1FFull <<  9)) | (b & (0x1FFull <<  0))) >>  0] <<  0;
-		h |= lookup3to2[((w & (0x1FFull << 18)) | (b & (0x1FFull <<  9))) >>  9] << 15;
-		h |= lookup3to2[((w & (0x1FFull << 27)) | (b & (0x1FFull << 18))) >> 18] << 30;
-		h |= lookup3to2[((w & (0x1FFull << 36)) | (b & (0x1FFull << 27))) >> 27] << 45;
+		h |= ((uint64_t)(lookup3to2[((w & (0x1FFull <<  9)) | (b & (0x1FFull <<  0))) >>  0])) <<  0;
+		h |= ((uint64_t)(lookup3to2[((w & (0x1FFull << 18)) | (b & (0x1FFull <<  9))) >>  9])) << 15;
+		h |= ((uint64_t)(lookup3to2[((w & (0x1FFull << 27)) | (b & (0x1FFull << 18))) >> 18])) << 30;
+		h |= ((uint64_t)(lookup3to2[((w & (0x1FFull << 36)) | (b & (0x1FFull << 27))) >> 27])) << 45;
 		return h;
 	}
